requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
gradio>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
librosa>=0.10.0
soundfile>=0.12.0
matplotlib>=3.7.0
moviepy>=1.0.3
pydub>=0.25.0
huggingface-hub>=0.19.0
insightface>=0.7.3
mediapipe>=0.10.0
tqdm>=4.65.0
scipy>=1.11.0
faster-whisper>=0.10.0
requests>=2.31.0
onnxruntime>=1.16.0
python-dotenv>=1.0.0
